requirements.txt
absl-py==2.3.1
actionlib==1.14.3
angles==1.9.14
astunparse==1.6.3
attrs==25.3.0
bondpy==1.8.7
cachetools==5.5.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
dynamic-reconfigure==1.7.6
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
gast==0.4.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
image-geometry==1.16.2
importlib_metadata==8.5.0
importlib_resources==6.4.5
interactive-markers==1.12.2
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.6
joblib==1.4.2
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
keras==2.13.1
kiwisolver==1.4.7
laser-geometry==1.6.8
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
message-filters==1.17.4
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyparsing==3.1.4
pyserial==3.5
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
python-dateutil==2.9.0.post0
python-qt-binding==0.4.6
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
requests==2.32.4
requests-oauthlib==2.0.0
resource-retriever==1.12.10
rosbag==1.17.4
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosgraph==1.17.4
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rsa==4.9.1
rviz==1.14.26
scikit-learn==1.3.2
scipy==1.10.1
sensor-msgs==1.13.2
shiboken6==6.6.3.1
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
sounddevice==0.5.2
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-cpu==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
threadpoolctl==3.5.0
topic-tools==1.17.4
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.3
xacro==1.14.20
zipp==3.20.2
